# docs/openapi_v9.yaml
openapi: 3.0.3
info:
  title: Plan Charge v9 API
  version: 1.0.0
  description: |
    OpenAPI spec for Plan Charge v9 backend (FastAPI) covering identity, planning, integrations, and reporting.
servers:
  - url: /api/v1
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Project:
      type: object
      required: [name]
      properties:
        id: { type: string, format: uuid }
        org_id: { type: string, format: uuid }
        key: { type: string }
        name: { type: string }
        status: { type: string, enum: [proposed, active, paused, done, cancelled] }
        priority: { type: integer }
        start_date: { type: string, format: date, nullable: true }
        end_date: { type: string, format: date, nullable: true }
        owner_id: { type: string, format: uuid, nullable: true }
        team_id: { type: string, format: uuid, nullable: true }
        tags:
          type: array
          items: { type: string }
...
    Task:
      type: object
      required: [project_id, title]
      properties:
        id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        epic_id: { type: string, format: uuid, nullable: true }
        title: { type: string }
        description: { type: string, nullable: true }
        status: { type: string, enum: [todo, in_progress, blocked, done, cancelled] }
        estimate_hours: { type: number, nullable: true }
        start_date: { type: string, format: date, nullable: true }
        due_date: { type: string, format: date, nullable: true }
        tags:
          type: array
          items: { type: string }
        order_index: { type: integer }
...
paths:
...
  /projects:
    get:
      tags: [Projects]
      summary: List projects
      responses:
        '200': { description: OK }
    post:
      tags: [Projects]
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Project' }
      responses:
        '201': { description: Created }
...
  /tasks:
    get:
      tags: [Tasks]
      summary: List tasks
      parameters:
        - name: project_id
          in: query
          schema: { type: string, format: uuid }
        - name: status
          in: query
          schema: { type: string }
        - name: assignee_person_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200': { description: OK }
    post:
      tags: [Tasks]
      summary: Create task
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Task' }
      responses:
        '201': { description: Created }
...